2020 Day 20
This was the hard day. Luckily, the problem was very similar to
the Drive ya nuts puzzle I solved a few years ago. Just with 144
pieces instead of 7, and 8 orientations instead of 6. This
solution is far from optimized. Apparently you can assume that
all edge pairs are unique, which would greatly reduce placement
time.