2020 Wrap-up
As always, I had a blast and learned a few things along the way.
Considering the rest of the year, this was a highlight of 2020.

I feel like this year was a bit easier than past years, but I
can't say how much of that is experience or actual difficulty.
To be honest, I think that my algorithm skills have faded a bit
after graduating. That constant practice in school really helps.

I found C# to be an okay language for this purpose, but definitely
felt like Python would have made my life easier several times.
That probably shows with my (over)use of LINQ.

I do intend to go back and do the previous two years' puzzles,
but after I take a small break. Let's hope my vacation goes better
than that poor elf's.