2022 Day 22 Improvement
Use circularly linked data structure instead of char indexes and
hash maps. This is an unusually tough thing to do with Rust
because of its strict enforcement of single ownership. But I had
to play around with Rc and RefCell before the year was up to get
the full Rust experience. Now that I have, I'm looking forward to
not using Rust anymore.