2023 Day 1
Starting the year off mildly annoyed by a poorly defined problem.
It makes no sense for lines to be interpreted differently when
checking for the last digit. The digits are still written RTL.
And yet, the solution expects "oneight" to be interprted as both
"1ight" and "on8" depending on which digit you're looking for.