2024 Day 24
Part 2 is my least favorite kind of problem, where you have to
visually inspect the input to find properties of the input not
specified in the problem that, if they weren't true, would make
the problem impossible to solve. I wrote some code that covers
the cases in my input, but it probably won't work for all inputs.